requirements_additions.txt
# Допълнителни зависимости за подобренията
python-dotenv>=1.0.1
Flask-WTF>=1.2.1
waitress>=3.0.0
openpyxl>=3.1.2
reportlab>=4.1.0
pywebview>=5.0
